ants

+++ validation.c
@@ -1,6 +1,6 @@
 #include "lem-in.h"
 
-void	valid_antcount(int antscount)
+void	valid_antcount(ssize_t antscount)
 {
 	if (antscount <= 0)
 		add_error("Invalid ants number");
